refactoring and comments in deploy script

+++ pantheon-deploy.js
@@ -1,3 +1,13 @@
+/**
+ * Deploy contents of ./static to Simplechart Dev Mode plugin on Pantheon
+ * SFTP username, password, host, and port can be specified in ./pantheon-config.json
+ * or passed from command line. Command line args will override JSON config args.
+ *
+ * If you just want to test the connection without transferring files, you can
+ * use `--test-connection` from the command line or `"testConnection": true` in JSON config
+ *
+ * node pantheon-deploy.js [--test-connection] [--username=XX] [--password=XX] [--host=XX] [--port=XX]
+ */
 const fs = require('fs');
 const path = require('path');
 const Client = require('ssh2-sftp-client');
@@ -5,20 +15,36 @@ const args = require('yargs').argv;
 
 const remoteJSPath = './code/wp-content/plugins/simplechart-dev-mode/js';
 
-function doDeploy(config) {
-  [
+/**
+ * Get CLI args relevant to SFTP connection
+ *
+ * @return {Object} key-value pairs of SFTP connection info
+ */
+function getConnectionArgs() {
+  return [
     'username',
     'password',
     'host',
     'port',
-  ].forEach((key) => {
+  ].reduce((acc, key) => {
     if (args[key]) {
-      config[key] = args[key].toString() || '';
+      acc[key] = args[key].toString();
     }
-  });
+  }, {});
+}
 
+/**
+ * Connect by SFTP and deploy static files
+ *
+ * @param {Object} config username, password, host, port for SFTP connection
+ */
+function doDeploy(config) {
   const sftp = new Client();
   sftp.connect(config).then(() => {
+    if (args['test-connection'] || config.testConnection) {
+      console.log(`Connected to ${config.host}; exiting`);
+      process.exit(0);
+    }
     console.log('Removing existing JS dir');
     return sftp.rmdir(remoteJSPath, true);
   }).then(() => {
@@ -34,6 +60,7 @@ function doDeploy(config) {
       ).then(() => {
         return sftp.list(remoteJSPath);
       }).then((list) => {
+        // Exit when all files are uploaded
         if (list.length === localFiles.length) {
           console.log('Uploaded files:');
           console.log(list.map(({ name }) => name));
@@ -43,11 +70,14 @@ function doDeploy(config) {
     });
   }).catch((err) => {
       console.log(err);
-      process.exit(0);
+      process.exit(1);
   });
 }
 
+/**
+ * Check for patheon-config.json then proceed with deploy
+ */
 fs.readFile(path.join(__dirname, 'pantheon-config.json'), 'utf8', (err, data) => {
   const config = ! err ? JSON.parse(data) : {};
-  doDeploy(config);
+  doDeploy(Object.assign(config, getConnectionArgs()));
 });
